# landing/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "inbox.dog API",
    "version": "1.0.0",
    "description": "API for managing email inbox access via OAuth. Create API keys, authenticate with Google, and manage credits for email operations."
  },
  "servers": [
    {
      "url": "https://inbox.dog",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/keys": {
      "post": {
        "summary": "Create API key",
        "description": "Creates a new API key with client credentials",
        "operationId": "createApiKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Optional name for the API key"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/keys/{client_id}": {
      "get": {
        "summary": "Get API key info",
        "description": "Retrieves information about an existing API key",
        "operationId": "getApiKey",
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The client ID of the API key"
          },
          {
            "name": "X-Client-Secret",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Client secret for authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "API key information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/oauth/authorize": {
      "get": {
        "summary": "Start OAuth flow",
        "description": "Initiates OAuth authorization flow and redirects to Google consent screen",
        "operationId": "authorizeOAuth",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your client ID"
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "URI to redirect to after authorization"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["email", "email:read", "email:send", "email:full"]
            },
            "description": "Requested scope for email access"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "State parameter for CSRF protection"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to Google consent screen"
          }
        }
      }
    },
    "/oauth/token": {
      "post": {
        "summary": "Exchange code for tokens OR refresh token",
        "description": "Exchange authorization code for access and refresh tokens, or use refresh token to get new access token",
        "operationId": "exchangeToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AuthorizationCodeRequest"
                  },
                  {
                    "$ref": "#/components/schemas/RefreshTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TokenResponseWithEmail"
                    },
                    {
                      "$ref": "#/components/schemas/TokenResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "Payment required (insufficient credits)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkout": {
      "post": {
        "summary": "Create Stripe checkout for credits",
        "description": "Creates a Stripe checkout session to purchase credits",
        "operationId": "createCheckout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Returns the health status of the service",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiKeyResponse": {
        "type": "object",
        "required": ["client_id", "client_secret", "name", "credits"],
        "properties": {
          "client_id": {
            "type": "string",
            "description": "The client ID for the API key"
          },
          "client_secret": {
            "type": "string",
            "description": "The client secret (only returned once)"
          },
          "name": {
            "type": "string",
            "description": "Name of the API key"
          },
          "credits": {
            "type": "number",
            "description": "Available credits"
          }
        }
      },
      "ApiKeyInfo": {
        "type": "object",
        "required": ["client_id", "name", "credits", "created_at"],
        "properties": {
          "client_id": {
            "type": "string",
            "description": "The client ID for the API key"
          },
          "name": {
            "type": "string",
            "description": "Name of the API key"
          },
          "credits": {
            "type": "number",
            "description": "Available credits"
          },
          "created_at": {
            "type": "number",
            "description": "Unix timestamp of when the key was created"
          }
        }
      },
      "AuthorizationCodeRequest": {
        "type": "object",
        "required": ["code", "client_id", "client_secret"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Authorization code from OAuth flow"
          },
          "client_id": {
            "type": "string",
            "description": "Your client ID"
          },
          "client_secret": {
            "type": "string",
            "description": "Your client secret"
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["grant_type", "refresh_token", "client_id", "client_secret"],
        "properties": {
          "grant_type": {
            "type": "string",
            "enum": ["refresh_token"],
            "description": "Must be 'refresh_token'"
          },
          "refresh_token": {
            "type": "string",
            "description": "The refresh token"
          },
          "client_id": {
            "type": "string",
            "description": "Your client ID"
          },
          "client_secret": {
            "type": "string",
            "description": "Your client secret"
          }
        }
      },
      "TokenResponseWithEmail": {
        "type": "object",
        "required": ["access_token", "refresh_token", "token_type", "expires_in", "email"],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The access token"
          },
          "refresh_token": {
            "type": "string",
            "description": "The refresh token"
          },
          "token_type": {
            "type": "string",
            "enum": ["Bearer"],
            "description": "Token type"
          },
          "expires_in": {
            "type": "integer",
            "description": "Token expiration time in seconds",
            "example": 3600
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address associated with the token"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "required": ["access_token", "token_type", "expires_in"],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The access token"
          },
          "token_type": {
            "type": "string",
            "enum": ["Bearer"],
            "description": "Token type"
          },
          "expires_in": {
            "type": "integer",
            "description": "Token expiration time in seconds",
            "example": 3600
          }
        }
      },
      "CheckoutRequest": {
        "type": "object",
        "required": ["client_id", "client_secret"],
        "properties": {
          "client_id": {
            "type": "string",
            "description": "Your client ID"
          },
          "client_secret": {
            "type": "string",
            "description": "Your client secret"
          },
          "credits": {
            "type": "number",
            "description": "Number of credits to purchase (optional)"
          }
        }
      },
      "CheckoutResponse": {
        "type": "object",
        "required": ["checkout_url", "session_id"],
        "properties": {
          "checkout_url": {
            "type": "string",
            "format": "uri",
            "description": "URL to redirect user to Stripe checkout"
          },
          "session_id": {
            "type": "string",
            "description": "Stripe checkout session ID"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": ["status", "service"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ok"],
            "description": "Health status"
          },
          "service": {
            "type": "string",
            "enum": ["inbox.dog-oauth"],
            "description": "Service name"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code"
              },
              "message": {
                "type": "string",
                "description": "Error message"
              },
              "action": {
                "type": "string",
                "description": "Suggested action to resolve the error"
              },
              "docs": {
                "type": "string",
                "format": "uri",
                "description": "Link to documentation"
              }
            }
          }
        }
      }
    }
  }
}